+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Changelog
 
 ## Version 0.0.1
-- Init
+- Backswing trait